+++ CompatibilityLib/common/src/main/java/com/elmakers/mine/bukkit/utility/CompatibilityConstants.java
@@ -17,6 +17,7 @@ public class CompatibilityConstants {
     public static final int FIREWORK_TYPE = 76;
     public static final UUID SKULL_UUID = UUID.fromString("3f599490-ca3e-49b5-8e75-78181ebf4232");
 
+    public static String LORE_WRAP_PREFIX = " ";
     public static int MAX_LORE_LENGTH = 24;
     public static int MAX_PROPERTY_DISPLAY_LENGTH = 50;