+++ CompatibilityLib/common/src/main/java/com/elmakers/mine/bukkit/utility/platform/CompatibilityUtils.java
@@ -63,7 +63,6 @@ import org.bukkit.util.BlockVector;
 import org.bukkit.util.Vector;
 import org.spigotmc.event.entity.EntityDismountEvent;
 
-import com.elmakers.mine.bukkit.api.magic.Messages;
 import com.elmakers.mine.bukkit.utility.BoundingBox;
 import com.elmakers.mine.bukkit.utility.DoorActionType;
 
@@ -253,8 +252,6 @@ public interface CompatibilityUtils {
     boolean sendActionBar(Player player, String message);
 
     void sendChatComponents(CommandSender sender, String containsJson);
-
-    void setMessages(Messages messages);
 
     float getDurability(Material material);